+++ packages/models/src/helpers/omnichannel/agentStatus.ts
@@ -1,0 +1,39 @@
+import { UserStatus } from '@rocket.chat/core-typings';
+import type { IUser } from '@rocket.chat/core-typings';
+import type { Filter } from 'mongodb';
+
+export const queryStatusAgentOnline = (extraFilters = {}, isLivechatEnabledWhenAgentIdle?: boolean): Filter<IUser> => ({
+	statusLivechat: 'available',
+	roles: 'livechat-agent',
+	// ignore deactivated users
+	active: true,
+	...(!isLivechatEnabledWhenAgentIdle && {
+		$or: [
+			{
+				status: {
+					$exists: true,
+					$ne: UserStatus.OFFLINE,
+				},
+				roles: {
+					$ne: 'bot',
+				},
+			},
+			{
+				roles: 'bot',
+			},
+		],
+	}),
+	...extraFilters,
+	...(isLivechatEnabledWhenAgentIdle === false && {
+		statusConnection: { $ne: 'away' },
+	}),
+});
+
+export const queryAvailableAgentsForSelection = (extraFilters = {}, isLivechatEnabledWhenAgentIdle?: boolean): Filter<IUser> => ({
+	...queryStatusAgentOnline(extraFilters, isLivechatEnabledWhenAgentIdle),
+	$and: [
+		{
+			$or: [{ agentLocked: { $exists: false } }, { agentLockedAt: { $lt: new Date(Date.now() - 5000) } }],
+		},
+	],
+});