+++ ee/packages/network-broker/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rocket.chat/network-broker
 
+## 0.2.31
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-services@0.13.1
+  </details>
+
 ## 0.2.30
 
 ### Patch Changes